backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
boto3==1.35.0
httpx==0.27.0
redis==5.1.0
Pillow==10.4.0
tweepy==4.14.0
python-dotenv==1.0.1
